alpha version after some refactoring. Created .\src\data_base.py. Refactored class DataBase in .\src\data_base.py, class MainScene, class HistoryScene, class SettingScene in app.py. Renamed some methods and functions due to the refactoring.
