+++ src/data_base.py
@@ -1,0 +1,48 @@
+
+import sqlite3 as sql
+
+
+class DataBase:
+    def __init__(self, db_name):
+        self.conn = sql.connect(db_name)
+        self.cur = self.conn.cursor()
+
+    def create_new_tables(self):
+        with self.conn:
+            self.cur.execute(""" CREATE TABLE  IF NOT EXISTS history (
+                                id INTEGER PRIMARY KEY NOT NULL,
+                                deposit INTEGER NOT NULL,
+                                date timestamp
+                                )
+                                """)
+
+        with self.conn:
+            self.cur.execute("""CREATE TABLE IF NOT EXISTS objective(
+                                id INTEGER PRIMARY KEY NOT NULL,
+                                objective INTEGER NOT NULL
+                                )""")
+
+    def insert_objective(self, obj):
+        with self.conn:
+            self.cur.execute(""" INSERT OR REPLACE INTO objective(id, objective) VALUES (?, ?)""", (1, float(obj)))
+
+    def get_objective(self):
+        self.cur.execute("""SELECT * FROM objective WHERE objective""")
+        return self.cur.fetchone()
+
+    def insert_history_log(self, deposit, date):
+        with self.conn:
+            self.cur.execute(""" INSERT OR REPLACE INTO history(deposit, date) VALUES (?, ?)""", (float(deposit), date))
+
+    def get_all_history_logs(self):
+        self.cur.execute("""SELECT * FROM history""")
+        return self.cur.fetchall()
+
+    def erase_history_log(self, id_num):
+        with self.conn:
+            self.cur.execute("""DELETE FROM history WHERE id=?""", (id_num,))
+
+    def erase_all_tables(self):
+        with self.conn:
+            self.cur.execute("""DELETE FROM objective""")
+            self.cur.execute("""DELETE FROM history""")
